requirements.txt
# requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.2.0
psycopg2-binary==2.9.7
redis==4.5.5
requests==2.31.0
PyPDF2==3.0.1
python-docx==0.8.11
groq==0.6.0
gunicorn==21.2.0
whitenoise==6.5.0
dj-database-url==1.3.0
django-redis==5.2.0
django-storages==1.13.2
boto3==1.28.62
python-multipart==0.0.6